SRE-697 reorganize application APIs
